Fix issue with $filter order of operations
When the filter for the tombstone preprocessor and the skip token
preprocessor was composed brackets were missing leading to
an incorrect evaluation of the $filter.

The repro required mulitiple key columns so that the paging generated a
multiple or terms.  On paged responses deleted entities were being
returned for non-delta queries.