Fix skiptokens with special characters.
The skiptokens were being double excaped.
e.g. a space ' ' would become %2520 instead of %20

Also '-' was being used delimit the skip token.
This is a common character in GUIDs and was breaking
the parsing of the skiptoken.  Nestle worked around the
issue, but putting the fix in to protect against any
future problems.

The delimiter is now ~~
In addition, any ~s in the key values will be changed to
~- so you can even use ~~ in the keys without any issues.